Migration Dalamud API15 + bump UmbraAPI 3000 → 4000
- Refactor des 4 handlers IChatGui.ChatMessage vers le nouveau modèle objet IHandleableChatMessage (ChatNameReplacement, ChatTargetSound, ChatEmoteHighlight, ChatProximityBlend)
- DalamudAsset.NotoSansJpMedium → NotoSansCjkMedium (asset renommé en API15)
- ImRaii.IEndObject → ImRaii.TabItemDisposable dans DataAnalysisUi et PlayerAnalysisUI
- DalamudUtilService: _lastZone passe de ushort à uint pour anticiper IClientState.TerritoryType → uint, et retrait du commentaire obsolète sur le fallback IClientState.LocalPlayer
- Bump submodule UmbraAPI vers ApiVersion 4000 pour forcer la mise à jour côté joueurs
- Bump version 2.5.1.4028 → 2.5.1.4029 + entrée changelog

+++ UmbraSync/Services/DalamudUtilService.cs
@@ -59,14 +59,12 @@ public class DalamudUtilService : IHostedService, IMediatorSubscriber
     private DateTime _delayedFrameworkUpdateCheck = DateTime.UtcNow;
     private string _lastGlobalBlockPlayer = string.Empty;
     private string _lastGlobalBlockReason = string.Empty;
-    private ushort _lastZone;
+    private uint _lastZone;
     private readonly Dictionary<string, PlayerCharacter> _playerCharas = new(StringComparer.Ordinal);
     private readonly List<string> _notUpdatedCharas = [];
     private bool _sentBetweenAreas;
     private static readonly Dictionary<uint, PlayerInfo> _playerInfoCache = new();
 
-    // Prefer IObjectTable.LocalPlayer when available (newer Dalamud),
-    // otherwise fall back to obsolete IClientState.LocalPlayer (suppressed) for compatibility.
     private IPlayerCharacter? LocalPlayer => _objectTable.LocalPlayer;
